requirements.txt
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.1
Flask-WTF==1.0.1
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
WTForms==3.0.1
